smp: Fix use after free in smp::invoke_on_others
The call sequence looks like this

smp::invoke_on_others(Func func)
 `- parallel_for_each([func = std::move(func)]
     `- smp::submit_to(func)

The func argument is first moved onto parallel_for_each's lambda then
the captured value is passed as an argument to the smp::submit_to. At
this point the func is considered as an lvalue reference on the value
that sits on the parallel_for_each's lambda capture. Respectively the
smp queue's async_work_item saves this reference onboard and the func
itself gets destroyed on exit from invoke_on_others.

Fix this by forcing copy of func when pushing it into submit_to. This
makes submit_to treat func as rvalue reference and save it (by moving)
on the async_work_item.

Since now copying happens add a concept requiring that the Func is
nothrow copyable as well.

fixes: #946

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>