files: Keep dioattr on fs_info
Now when the fs_info travels all the way down the posix_file_impl
constructor, it's possible to keep the XFS' dioattr data on it.
The member is made optional to work on non-XFS.

While at it -- rename the query_dma_alignment into configure_ as
it doesn't really query anything any longer.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>