files: Don't ignore FS info for read-only files
When a file is opened with O_RDONLY flag the make_file_impl treats
it as a directory and opens with hardcoded 4k block-size and nowait
flag turned off.

It seems that regular files, whatever the mode is, should go through
the same procedure.

tests: unit(dev), scylla.unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210923055832.30051-1-xemul@scylladb.com>