file: keep hint pointer alive when calling fcntl()
per fcntl(2), we can set/get file read/write hints using fcntl(),
and the commands like F_GET_RW_HINT takes parameter of `uint64_t*`.
we should keep the pointer alive when calling fcntl() for
setting/getting read/write hints. before this change, the pointer to a
local variable is passed to a function which schedule an async call to
fcntl(), so there is no guarantee that the pointer is still valid when
the async call (continuation) is scheduled.

in this change, do_with() is used to keep the uint64_t "hint" alive
when calling fcntl(command, &hint).

this change should addresses the test failure of Seastar.unit.file_io
in debug mode.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20211004095529.958730-1-tchaikov@gmail.com>