memory: always allocate buf using "malloc" for non reactor
before this change, there is chance that Seastar reallocate a memory
buffer allocated by the c runtime even before "original_malloc_func"
is initialized. the runtime might want to allocate memory for some TLS
variables or just for itself, like env variables.

but we might want to call functions like setenv() from a reactor thread.
and setenv() could reallocates the memory buffer previously allocated by
the "main" thread. in this case, cpu_pages::shrink() panics if it is
asked to resize a memory chunk allocated by another reactor thread.

in this change,  __libc_malloc and friends are not overridden anymore.
instead of using dlsym(RTLD_NEXT) for original_malloc_func, use
__libc_malloc.

the same applies to other malloc related functions.

because musl libc does not define __malloc_trim(), __libc_memalign()
or __malloc_usable_size() yet, see
https://git.musl-libc.org/cgit/musl/tree/src/include/stdlib.h .
so the corresponding overrides just call the underlying function
only if it is available, and call the dlsym()'ed function as a fallback.
this is based on the assumption that these fallback functions is not
called by dlsym().

the solution was originally proposed by Avi Kivity <avi@scylladb.com>

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20211004175823.1750511-1-tchaikov@gmail.com>