Scheduler: Add support for scheduling group count configuration
Here we add support for setting the number of max scheduling groups
at compile time.
A support for this has been added to the CMake recipe in the form
of a parameter: Seastar_SCHEDULING_GROUPS_COUNT which translates
to a target definition SEASTAR_SCHEDULING_GROUPS_COUNT which is later
used in the code to set the max_scheduling_group return value.
The support has also been added to the configure.py script with
the paremater --scheduling-groups-count.

Tests:
1. Tested with the default size which is 16
2. Tested with a number which is not power of 2 (17)
3. Tested with a non number as parameter (Produces CMake informative
   error)
4. Tested with a number greater than 64 (sizeof(unsigned long) *
   CHAR_BITS) - A pre-existing static assert fails.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>