cpu_stall_detector: handle wraparounds in Linux perf_event ring buffer
Records in the Linux perf_event ring buffer can wrap, with the first
part of the record discontiguous with the other part.

Rather than try to make the data contiguous, we observe that the
records aren't really structs - they are variable size and most
of the data is optional. So read them incrementally in 64-bit
units (which can't wrap).

We add a data_area_reader class to encapsulate the details, and
rewrite the reporting code to use the reader. We also report all
records, not just the last one, in case they provide extra
information.
Message-Id: <20211008165705.80998-1-avi@scylladb.com>