queue: mark methods depending on std::queue noexcept
Although the respective std::queue methods are not
specified as noexcept they have no reason to throw.
For example, since we use seastar::circular_buffer,
std::queue::pop just calls circular_buffer::pop_front
and the latter is specified as noexcept.

To simplify marking of methods that rely on those as noexcept,
mark the low-level methods as noexcept.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>