file: ~append_challenged_posix_file: allow close to fail truncate
It is only critical that the ~append_challenged_posix_file
queue is empty when destroyed, to prevent use-after-free.

If close failed to truncate the file to the logical_size
we should allow destroying the file gracefully, since
the user is required to call flush() before close() to
guarantee stable storage semantics.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>