queue: require element type to be nothrow move-constructible
As it is required to be so to be returned as future<T>.
The static_assert for that in future_base goes off
in this case, but the compiler error is quite obscure
so lock that requirement at the queue level as well.
It will be taken advantage of in the following patches.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>