build: Include dpdk as a single object in libseastar.a
this change brings 4bd84425084e25bc957c5693b7a299739b7f039f back.
which was previously reverted by
76c87384dc8866aae603a4b598ff0adc8c9bac5a.

76c87384dc8866aae603a4b598ff0adc8c9bac5a intended to bring the DPDK
dependencies as an interface library instead of *including* the DPDK
libraries in libseastar.a as a relocatable output file. after
76c87384dc8866aae603a4b598ff0adc8c9bac5a, in addition to linking against
libseastar.a, we have to link against DPDK libraries when linking a
Seastar application when DPDK is enabled. this renders the
redistribution of the Seastar library more complicated. because we
always build our own dpdk, without including dpdk into libseastar.a,
we have to redistribute the DPDK libraries along with libseastar.

also, the rationales explained in the commit message of
4bd84425084e25bc957c5693b7a299739b7f039f still hold.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20211018172240.444417-1-tchaikov@gmail.com>