file: append_challenged_posix_file: close: report error from truncate
Current close makes a best effort to truncate the
file to the logical size, but that may fail.
We currently swallow the error silently, so expose
it so it can be logged 9and ignored).

Define a new private method - truncate_to_logical_size
that can be reused by both flush() and close().

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>