io_tester: Enhance thinker
Right now the think time is the amount of time a fiber sleeps after
each request. Enhance this by making it possible to control the whole
priority class sleeping periods.

The new parameter is 2 numbers -- time to keep submitting request and
time to idle. Both, as was said above, as class-wide, not fiber wide
to ensure the full class is idling at times, not just individual
fibers. The latter may result in partially overlapping sleeps and
never really idling class.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>