fair_queue: Hide priority_class_data
Now the fair_queue's priority_class(_data) is internal to fair_queue
code. Hide it inside the fair_queue class (and move to fair_queue.cc)
like it was done with the io_queue's pclass data in the previous
patch.

Rename the guy to have the same name as its io_queue peer while at it.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>