io_queue: Keep several fair queues onboard
Each fair queue represents an independent stream. In this series we need
at most two, so keep them in a small_vector<> with 2 inline entries.

A vector, rahter than two explicit fair_queues is selected to make it
possible to reference the fair_queue from stream_id by plain indexing.

Another trickery here is that fair_queues can sit directly in a vector,
because the are move-able (at the early stage -- 100% moveable).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>