util/log-cli: expose parse_map_associations()
Moved from util/program-options.cc where it was a private static
function. Refactored to take a consume function, as its next user (added
in the next patch) will want to consume the entries differently.
This code is not performance sensitive so better to use a std::function
here, instead of templates.