on_internal_error: set_abort_on_internal_error: return current state
Return the current abort_on_internal_error state
when changing it via set_abort_on_internal_error.

This is useful for tests that wish to disable
abort_on_internal_error temporarily, and in particular
use it in scoped_no_abort_on_internal_error
so it can be nested, or performed under a global
false setting via the command line.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>
Message-Id: <20211115125848.3961766-1-bhalevy@scylladb.com>