fair_queue: Dont export head value from group
The head and tail rovers of the group represents the charged capacity.
Nowadays this capacity is measured in terms of tickets, but it's going
to change. This patch hides the head exact type and value inside the
group itself, so that the caller (fair queue) just checks if the
capacity is available or not without knowing what it actually is.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>