net: bring local_address() to seastar::connected_socket.
This patch extends the interface of `connected_socket` to provide
clients with a way to obtain the socket's local address. It's built
on top of the already existing low-level infrastructure.

The new method is necessary (no other way AFAIK) to obtain address
of a `connected_socket` returned from `seastar::connect()` as -- in
contrast to `seastar::accept()` -- there is no a counterpart of
`accept_result` which would hold `socket_address`.

However, the new facility seems useful also when handling `accepted`
instances as it provides same interface which also resembles
`seastar::server_socket`.

The new feature is needed in crimson-osd and Rook.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>
Message-Id: <20211125161113.107738-1-rzarzyns@redhat.com>