smp: destroy_smp_service_group: verify smp_service_group id
Verify that the smp_service_group id to be destroyed
is valid, otherwise we end up with clearing a vector
in an arbitrary position in memory.

To protect against use of stale smp_service_group:s,
add a version field in debug mode and verify it too
in destroy_smp_service_group.

We keep only 16 bits for the group id in debug mode so
limit the total number of smp service groups to 2^16 (65536).
This is a reasonable limit for all modes.

call on_fatal_internal_error if the ssg id is invalid.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>