io_queue: Configure queue in terms of blocks
Currently io-queue configuration uses "bytes". When the io-queue
assigns tickets to requests it converts request lenghts into
512-bytes blocks. Respectively, the fair-queue is configured in
terms of blocks.

It's a bit messy and creates additional difficulties in the
future patching. Clean this place -- blocks appear at the IO level
and it's up to reactor configuration code to convert io-properites
from bytes to blocks.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>