log: Silence logger when logging
When logger is doing its job it can deadlock in syslog() call. The
latter calls malloc() which in turn can warn large allocation (or
error the failed one). Logging from inside of syslog() would call
syslog() again and it will self-deadlock on internal mutex.

Prevent this (and any future recursion) early with the help of a
ephemeral log-level called "silent" whose numerical value is lower
than any other level so that is_enabled() check would return false.
This helps to avoid extra if's on hot paths.

fixes: #980

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20211206190755.26560-1-xemul@scylladb.com>