core/program_options: preserve defaultness of CLI arguments
Currently the visitor which extracts CLI arguments from
boost::program_options::variable_map and propagates them to the new
seastar::program_options descriptor structure doesn't consider
defaultness of said arguments and if the value is present at all, it
will set it will propagate default values as non-default ones.
One visible effect of this was that --overprovisioned was ignored when
it came to not pinning shards to CPU cores because whether to pin or not
was decided by checking `overprovisioned &&
thread_affinity.defaulted()`, which due to default being forced to true
for all values with a default value, always evaluated to true.
Fix by not setting a value to variables either missing on the CLI or
being defaulted, leave these variables defaulted.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20211208075425.106982-1-bdenes@scylladb.com>