fair_queue: Linearize accumulators
Currently the per-class accumulated costs are calculated like

  cost = request_cost / pc.shares * exp((time - time_0) / tau)

The exponential add-on is aimed at solving re-activation problem,
i.e. when a class gets back from being idle it can have its
accumulator very low thus monopolizing the queue for the period
it was idling. Exponential costs prevent this from happening.

At the same time exponential costs are not predictable when it
comes to rate-limited input. E.g. if some class pushes requests
at specific rate the way it "shares" its accumulator with the
unbound class is not

   a.shares / b.shares

The formula is more like

  log_A ( a.shares / b.shares * A + 1 ) - 1

where A is a parameter depending on the submission rate like

  A = exp(rps_period / tau)

This makes capacity sharing hard to analyze, all the more so the
rps-period is not constant.

Instead, replace the accumulator with just

  request_cost / pc.shares

values adding up to one another.

The re-activation is solved by tracking the "minimal" accumulator
value served so far. When re-activating the class is allowed to
have its accumulator lower than the minimal one, but the difference
shouldn't make the class monopolize the queue for more than
tau duration.

The above change breaks one of the fair_queue test cases that
expects that the re-activated "priority" expires in a timely manner.
In the new model the new class should actually consume this
priority before it vanishes.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>