util: closeable/stoppable: Introduce cancel()
Useful in situations which need to guard for close() in a block before
passing the closeable object further, moving the responsibility for
closing to a different scope. Passsing the guard together with the
object is another way to solve that, but more cumbersome.

There is a precedent of this API for such guards:

  deferred_action::cancel()
Message-Id: <20211220155829.686969-1-tgrabiec@scylladb.com>