reactor: improve lowres_clock, lowres_system_clock granularity
lowres_clock and lowres_system_clock are faster versions of
std::chrono::steady_clock and std::chrono::system_clock, trading
accuracy for lower overhead. They do this by sampling the time
every 10ms and using that.

Unfortunately, this leads to unintended coordination: if many
running operations set a timer for (say) 1 second, the timers
will all expire at 10ms boundaries. The system will be flooded
with expirations on these 10ms intervals, and have less work
at other times.

This usually isn't a problem, because most timers do not expire
at all - they are used for timeouts (that typically are cancelled
when an operation completes successfully). But io_tester does
use expiring timers in order to pace request issue rate, and
suffers from this coordination problem (especially as it leads
to alternating between flooding and idling the disk).

Using high resolution timers isn't a good option, because they
are too high resolution - sacrificing performance for accuracy.
They do this by preempting the running task. This involves multiple
system calls and polling all event sources.

The solution is to improve lowres timer resolution to task quota
(typically 0.5ms) granualarity. This is the best-case scheduling
granualaity for I/O anyway, so better accuracy than that is wasted.
It turns out to be simple to achieve: every time we start a task
quota period we refresh the clocks from source. This is less
expensive than a system call (of which we do several at that time),
and more expensive that loading a thread-local variable (which we
do in the implementation of now()), so any performance loss is
negligible.

We also refresh the clocks when expiring timers, to make sure they
are expired in a timely manner.

Sleep mode happens to not require any change: we already maintain
the closest timer expiration and already set a high-resolution wakeup
for that time.

Unfortunately, this change can break user code if it assumes
lowres_clock::duration is milliseconds instead of using the exposed
type. I consider it a user bug, and so we don't need to provide an
API level for it (although it is unfortunate to trigger such
breakage).