fair_queue, io_queue: Configure replenish threshold from minimal IO request
The threshold in question controls how small can a duration be
to actually be converted into the capacity. If the resuling
capacity is smaller than this value replenisher lets more ticks
to pass, otherwise it puts more tokens into a bucket and updates
the shared replenishing timestamp.

Having this value small has several effects. First, the smaller it
is the more frequent access to shared rovers happens. Next, the
smaller it is, the more frequently tokens get into the bucket. The
latter, however, has a side-effect -- since the tokens are not
"generated" but are "moved" from another bucket, if that latter
bucket has not enough of them the converted duration is lost for
good, more tokens will be replenished after more time. So keeping
this value small should come with rare enough replenish rate not
to lose tokens too hard.

This patch increases the threshold to be as small as the smallest
capacity request a queue can have. This information is available
on the IO-queue level -- the smallest ticket comes from the lightest
request of one block.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>