fair_queue: Merge cost/shares capacities
Currently there are two tickets on the fair-group used to normalize
per-request tickets -- the cost_capacity and the shares_capacity. The
former one is used to evaluate how much a request contributes to the
disk run-time capacity (a.k.a. the tokens in the buckets), while the
former one is used to calculate the accumulator -- a counter used to
balance between classes according to their shares. The shares ticket
is exactly latency_goal[us] times larger than the cost one.

It's better to remove the shares ticket and use the cost one in all
the calculations. Note, that the accumulator values instantly become
~1k times less, but for now it's fine -- the accumulator is (still)
a floating-point number.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>