fair_queue: Treat accumulator as capacity
As was described in the previous patch -- capacity is the number of
tokens a request contributes to the bucket, while the accumulator is
the number of units a request contributes to its class. The latter
value is numerically equal to the former one so it's natural to use
these two interchangeably. All the more so the accumulated value on
a class represents the number of tokens consumed by this class in the
rate-limiting formula.

However, the accumulated values are the tokens divided by the number
of shares. For unconfigured queue the smallest "cost" of a request is
~1900 (the 4GB/s disk has this value ~50000), so large shares values
may produce something less than 1 and converting it to integer may
produce 0 which, in turn, means a class makes no progress and
monopolizes (locks) the queue. So make sure the division doesn't
produce 0 by chance.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>