io_queue: Move max blocks count onto group
This value is used to calculate request accumulator cost and to check if
the request length is too large. The former usage is about to be replaced
with rate-limter token cost, but it's still needed to know the maximum
length of the request, so keep the limit as explicitly separate param.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>