io_tester: Support /dev/null as storage
This is very useful to check CPU-bound behavior of reactor + scheduler
without disk messing with it.

Other than allowing to open the /dev/null itself, a trick is taken to
simulate that read() reported requested size, not 0 as /dev/null does.
That's for the statistics to look as if it was on real disk.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>