io_queue: Auto-increase the io-latency goal
It may happen that the disk iops/bandwidth are too low for even a
single request to be served within the scheduler tick. In this case
the latency-goal value should be increased.

In current scheduler implementation this means checking the token
bucket limit against the replenisher threshold. The limit means how
many requests can be submitted within the latency goal, while the
replenish threshold is the number of tokens needed for the minimal
request. If the limit is below the threshold then even the smallest
request doesn't fit, and the goal is too low. And since it's better
if seastar "just works" even if the configuration is not correct,
the first change here is to set the limit to be at least N times
larger than the threshold to let at least N minimal requests in.

N is selected to be 3, because the limit value is used to get the
maximum request _length_ (in bytes), and 1.5K smells somewhat better
than 512 bytes ... Oh, well.

Finally, the io-queue checks if the latency goal was increased by
the fair-queue, and if it was the warning is printed with the newly
set-up value.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>