reactor: Fix off-by-end-of-line misprint in legacy configuration
When the --max-io-requests is specified the rate-limiter assumption
should be that this is the number of requests approx. per-millisecond.
Then the std::min trim is applied on it not to set the request rate
larger than the amount of simultaneously executing AIOs in the kernel.
Also approx(imately).

The misprint was that the trim was applied to -per-second value, while
should have been to the initial -per-ms one. The result is usually 1k
times slower rate-limiter configuration than it should.

fixes #1006

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220125140718.31979-1-xemul@scylladb.com>