queue: pop: assert that the queue is not empty
Popping an element from an empty queue
leads to undefined behavior as we return _q.front()
that might have been a valid element in the past
that is now either used after free, or if it's e.g.
a shared ptr, may result in freeing it while in use
since its reference count is not incremented.

This change adds an assert that will turn the
undefined behavior into a crash to help find
these issue earlier and generate a coredump
since, as seen in e.g. scylladb/scylla#9955,
these kind of issues can lurk for a long time
and are very hard to debug.

Test: unit(dev), queue_test(debug)

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>
Message-Id: <20220126080847.977186-1-bhalevy@scylladb.com>