future: fix make_exception_future(const exception_ptr&&)
In commit 31a8e65551e36c416daab6487b54f7ea395296ff, the overload
make_exception_future(std::exception_ptr) was split - for performance
reasons - into three separate overloads:

        make_exception_future(exception_ptr&&)
        make_exception_future(const exception_ptr&)
        make_exception_future(exception_ptr&)

But a fourth overload has been missing:
        make_exception_future(const exception_ptr&&)

While an rvalue reference to a const object is not particularly useful,
it can happen unintentionally when a non-mutable lambda (which makes all
its captures const) does std::move() on a captured variable.

Because this fourth overload is missing, make_exception_future() will
fall back to its "default" implementation which puts the given object
inside an exception (using std::make_exception_ptr()), and this is the
wrong thing to do for an exception_ptr - const or not.

Besides the fix of adding the missing overload, this patch also includes
a new unit test for make_exception_future() which exercises the different
overload of make_exception_future, and fails one check before this patch
and passes with the patch.

Fixes #1010.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20220130202403.1206131-1-nyh@scylladb.com>