io_queue: Keep max lengths on board
When constructed the io_group calculates the maximum ticket value
acceptable by fair queue code. This math loses a bit of precision
when converts read lengths and write lengths into a single ticket
value. Things get simpler if limiting request lengths with the
lenghts per-se, not with the ticket value.

As the first step -- keep the maximum langhts for both IO directions
on the fair group.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>