io_queue: Introduce make_ticket() helper
There are two (actually three) places that calculate ticket value
from a request -- the request_fq_ticket and io_group constructor.
The third place is implicitly present in the former method and
makes a ticket for an oversized request.

This patch unifies both places and patches the limit checker to
work with maximum lengths, not maximum ticket size.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>