perftune.py: Introduce a framework to define deprecated parameters
Add a simple way to define a deprecated parameter.
No need to clean the code (although highly recommended).

All you need to do is to add a deprecated parameter definition
in filter_deprecated_args() and the framework will make the rest of
the script code "think" that this parameter hasn't been provided
even if it was provided.

And if you would be provided a warning is going to printed
to STDERR.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>