core/app_template: print debug warning to std::cerr
app_template::run_deprecated() prints a warning when running a debug
build. This is currently printed to stdout, which is inconsistent with
seastar logging (which by default is written to stderr) and more
importantly can interfere with any output the seastar application wants
to write to stdout.
Print it to stderr instead to fix this.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20220209095924.182979-1-bdenes@scylladb.com>