log: Fix silencer to be shard-local and logger-global
There are two issues with the log silencer.

First, it shuts up logging on all shards, while the intention was to
avoid reentrances into syslog by one shard. This is because the _level
field is shared between shards and setting it to "silent" is visible
everywhere.

Second, it doesn't fully solve the original problem. Lowering the level
on a logger will prevent it from self-reentrance, but it doesn't keep
different loggers from nesting into each other. Instead of silencing
a logger it should stop the whole shard from logging.

Said that, the fix is in introducing the thread-local "be quiet" bit
that would make all loggers stop logging, but only on a specific shard.
This new bit is +1 condition, but it sits outside of the builtin-expect
and gets into the slow-path of the logging code.

fixes: #1012
tests: unit(dev), ioinfo(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220210141845.25085-1-xemul@scylladb.com>