reactor, file: Fix block device detection
When a block device ioqueue is configured it's assigned a dev-id
from the "lower" device, i.e. the one on which a filesystem that
owns the former device file resides. This is not correct, device
files should have their own numbers.

This fix in reactor has the peer bug in file opening code -- when
a block device is opened it's given the same "lower" device number
using which the io-queue is selected. This place should use the
st_rdev as well.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>