Add abortable_fifo container
The container implement a fifo where each element can be associated with
its own abort source and can be removed from the fifo (out of order)
by triggering the abort source.

This is mostly a c&p from expiring_fifo with all its gotchas and
limitation and it purpose is to replace the expiring_fifo in the semaphore
and shared_future classes. It is put into internal namespace to avoid
general usage since it is not full-fledged container, but rather tailored
for its use case in those classes.