semaphore: make semaphore::wait() abortable
Provide semaphore::wait() variant that takes abort_source as a parameter
and can be aborted through it.

It is needed to make individual waits abortable as opposite to entire
semaphore to allow individual fibers to be aborted separately. This can
be useful for instance in systems that serve long running requests and
may want to abort some (but not all) because of a resource shortage or
because a client that requested the work no longer waits for a result.