shared_future, shared_promise: allow aborting wait for shared_future.
Add shared_promise::get_shared_future() variant that takes abort_source
as a parameter and can be aborted through it.

It is needed to make individual waiters abortable to allow fibers to
be aborted separately. This can be useful for instance in systems that
serve long running requests and may want to abort some (but not all)
because of a resource shortage or because a client that requested the
work no longer waits for a result.