tests: coroutines_test: Check scheduling group set with switch_to() is inherited and restored
The scheduling group set with coroutine::switch_to() must be inherited
from threads, newer coroutines frames, and the previous scheduling
group will be automatically restored once the coroutine frame, which
established the new sg, goes out of scope. Let's verify all that with
an unit test to guarantee we won't have regressions in the
behavior, which would break the applications that relied on it.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20220301130334.19643-1-raphaelsc@scylladb.com>