semaphore: expiry_handler: distinguish between broken_semaphore and abort_requested_exception
Distiguish between an exception generated via semaphore::broken()
vs. a particular wait() being abort via its abort_source.

In the first case, the exception is kept in semaphore::_ex
so just get it from there (by default it will use
the ExceptionFactory broken() excception, if available, or
semaphore::broken_semaphore otherwise.

If there's no semaphore-global exception set, it means
that expiry_handler was called on a particular entry.
Generate abort_requested_exception so that the caller of wait()
or get_units() would be able to distinguish between the cases.

Extend the existing exception factories with an aborted() exception
for representing this case.

Test: semaphore_test

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>