websocket: revamp shutdown routines
This commit makes the connections react to errors like incorrect
HTTP upgrade requests by shutting the connection down,
which is a reasonable thing to do.
Also, the gating mechanism copied from the http server does not
make the same amount of sense in case of websocket handlers,
because a websocket handler can be an indefinite stream that
can effectively block the gate forever until the client keeps
sending data. Thus, the gate is removed and the connections are
simply closed when the server stops.