coroutines: add parallel_for_each
This patch adds support for coroutine::parallel_for_each
that can be used in coroutines instead of
co_await seastar::parallel_for_each.

This implementation minimizes memory allocation
by deferring the allocation of a vector of futures
to wait on, till any of the futures is unavailable,
similar to the legacy parallel_for_each implementation.

Also it waits on the unavailable futures from back to front,
similar to to legacy implementation, to minimize
the number of callbacks required.

The advantage over the legacy implementation is
that the parallel_for_each object itself is a
seastar task that is used to wait on unavailable
futures with no need to allocate a parallel_for_each_state.

This is possible since coroutine::parallel_for_each
is defined as [[nodiscard]] and the caller must co_await
for it, making sure it remains valid until all the futures
it's waiting on are resolved.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>
Message-Id: <20220317153144.76189-1-bhalevy@scylladb.com>