coroutine: add a non-preemptive co_await flavor
It might be useful to be able to assume that co_awaiting on a ready
future will not yield. In order to do that, one can now explicitly
ask for immediately following a ready future with the following
construct:
  co_await coroutine::no_preempt_check(my_ready_future);

This commit comes with a unit test which is a blatant copy-paste
of the previous preemption test, but with the final condition reversed,
as well as a short paragraph in the tutorial.

Message-Id: <7608eba1c88ee5ca41b7bc9e2a2a898e05f65e81.1647534476.git.sarna@scylladb.com>