condition_variable: restore move-constructability.
This patch restores the move-constructability of
the `seastar::condition_variable` which is crucial
when e.g. a user of it needs to be passed to
`seastar::do_with()`. Lost of this ability was
the root cause of the following FTBFS in crimson-osd
(moving ctor has been manually added to `Throttle`
to illustrate the issue in a more concise way):

```
In file included from ../src/crimson/net/SocketConnection.h:20,
                 from ../src/msg/Message.h:39,
                 from ../src/auth/cephx/CephxSessionHandler.cc:23:
../src/crimson/common/throttle.h: In constructor ‘crimson::common::Throttle::Throttle(crimson::common::Throttle&&)’:
../src/crimson/common/throttle.h:26:49: error: use of deleted function ‘seastar::condition_variable::condition_variable(const seastar::condition_variable&)’
   26 |     : on_free_slots(std::move(rhs.on_free_slots)) {
      |                                                 ^
```

The problem has been introduced in e74a93c3934b2c8d360468bcdc03e2651dc1020d.

Signed-off-by: Radosław Zarzyński <rzarzyns@redhat.com>
Message-Id: <20220318132552.1087723-1-rzarzyns@redhat.com>