core/circular_buffer: fix iterator comparison for wrapped-around indexes
Indexes into a circular buffer can wrap-around (under or overflow) by
design. The code knows about this and it is handled correctly. One area
which managed to shelter a related bug undetected is iterator
comparison. Iterators are just thin wrappers around an index. The
comparison operators simply compare the two indexes. This breaks when
one of the two compared iterators wrapped around but the other didn't.
To fix, we compare offsets compared to the begin iterator.

A unit test reproducing the problem is also added.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20220325071753.42721-1-bdenes@scylladb.com>