core/circular_buffer: fix iterator comparison for wrapped-around indexes
Indexes into a circular buffer can wrap-around (under or overflow) by
design. The code knows about this and it is handled correctly. One area
which managed to shelter a related bug undetected is iterator
comparison. Iterators are just thin wrappers around an index. The
comparison operators simply compare the two indexes. This breaks when
one of the two compared iterators wrapped around but the other didn't.

The consequences of this is very severe: erase() can return early,
failing to call the destructors of some of the contained elements. If
the elements are RAII objects, the resource they are guarding will be
leaked.

To fix, we use operator- to calculate the difference between the
compared indexes, then compare this to 0.
A unit test is added which stresses the comparison running a random
sequence of push front, push back and erase, comparing all iterator
pairs at each iteration.
I have confirmed that it reproduces any problems with the comparison
both before and after the previous (now reverted) incomplete fix.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20220328121003.74925-1-bdenes@scylladb.com>