metrics: Introduce REAL_COUNTER
The new metrics type is internal and is only needed to tell the
prometheus exporter that the value in question is real number, not
integer, so when rendering the fraction part should not be forgotten.
This doesn't make any difference for the prometheus itself, it
doesn't tell integers from reals and handles everything as floating
point numbers. For seastar this trick is used not to perform extra
work -- formatting the floating point number takes more cpu cycles,
as well as an integer casted to float can be rendered as x.9999999999
if not being lucky enough.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>