condition_variable: fix cv.signal with active "when" wait would switch fiber
The awaiter type for "when" waiters would immediately issue a
coroutine_handle::resume() when signalled. This is not right, because
this effectively switches out the current fiber in favour of the
waiting one.

The correct approach is of course to implement task, and schedule
the awaiter to be resumed whenever reactor feels fair.
Message-Id: <20220330015657.18535-1-calle@scylladb.com>