memory: allow user to select allocator to be used at runtime
This option is targeted at small utility/tool applications, which don't
want to use the seastar allocator, as they don't want to commit to a
fixed amount of memory to be used. Any amount they might commit to would
be both too small and too large at the same time: we want to let them
consume as much as they need and only as much instead.
This option is not exposed on the command line, it can only be set from
code. This is intentional: the application code has to be well prepared
for the choice of allocator, this is not something a user can select at
whim on the command line, this is a decision that only developers can
make (with care).