rpc_tester: Add dummy client jobs
This adds a list (well, a vector) of abstract jobs to context.
Only client-owning context maintains this list, server has it
empty. When being ran, the context runs all its jobs in parallel.

Jobs description comes in YAML config. For now it only contains
job name and type. For now only the 'rpc' job type is supported,
but in the future there should appear the 'cpu'hog job, and
maybe some others.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>