coroutine/exception: make it work with co_await
The seastar::coroutine::exception wrapper was introduced in order to
return exceptions from coroutines without throwing. A wrapped exception
is supposed to be co_returned from a coroutine.

Unfortunately, due to limitations of the language, coroutines returning
a void future are not allowed to `co_return anything;`,
only `co_return;`, which prevents usage of the wrapper in coroutines
returning void futures.

This commit introduces the possibility of using the wrapper with
co_await. Co_awaiting a wrapper has the same effect as co_returning it,
but unlike co_return it is also legal to do in coroutines returning a
void future.

A helper function seastar::coroutine::return_exception is introduced
which works similarly to seastar::coroutine::make_exception but has a
name which expresses the intent better when used along with co_await.

Closes #1038