rpc: make it possible to silence connection errors
The object returned by `rpc::make_client()` (`struct shelper` defined
inside `send_helper`) now contains a member `send_options opts` which
can be used to configure the object before sending RPCs using its
`operator()()`.

We introduce an option `report_connection_error` which specifies whether
the RPCs sent using this object are interested in Seastar logging a
connection error, if such an error occurs. Seastar will then report
a connection error if and only if:
- there was at least one one-way RPC sent on this connection which
  specified report_connection_error=true,
- or there was at least one two-way RPC sent on this connection which
  specified report_connection_error=true and didn't return yet (cancel
  and timeout also count as returning).

Note that this does silence the errors in some cases, even if everyone
specifies report_connection_error=true (which is the default for
backwards compatibility). Namely, if a remote connects with us and sends
us RPCs, even if we respond to these RPCs, but we don't send any RPCs
ourselves (or send only RPCs with report_connection_error=false), and an
error occurs on this connection, we won't report it. That is, we
report a connection error only if *we* started RPCs on this connection.
(This behavior can be changed easily if needed in the future, by
changing one bool from `false` to `true`.)

The change only applies to non-stream connections.
Stream connections cannot be configured in this way and always report
connection errors.