rpc: lower log level for 'failed to connect' errors
Some distributed applications want to try contacting remote endpoints
periodically even when they think that those endpoints are down. For
example, the Scylla gossip module tries to contact remotes every second,
even if some of them are marked down by the failure detector. This
causes an unwanted spam of error messages in the logs.

Lower the log level for when the RPC module fails to establish a
connection to `debug`. It will still report a message on `error` level
if an error occurs on an already established connection.