reactor: Always limit configuration with active cpumask
When seastar app is run under the e.g. taskset shell command it
effectively ignores the enforced cpumask and pins its thread to any
core (well, to whatever core is allowed by current cgroup which is
pretty much often is "any").

Fix this by restricting seastar with current cpumask always. In case
the --cpus option is provided, check the selected cpus against this
current set like it was done against cgroup cpuset (it's backward
compatible because current cpuset is the subset of the cgroup one).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>