cmake: add FindValgrind.cmake
valgrind/valgrind.h is included in couple places in the source tree,
so we can use the macro of "RUNNING_ON_VALGRIND" for checking if
we are using on valgrind(s). this macro is a header-only one, and
it does not reference any symbols exposed by valgrind libraries.

in this change,

* `FindValgrind.cmake` is added to find Valgrind, so cmake fails
  early with hopefully more visible errors if "valgrind/valgrind.h"
  is not found.
* let the targets which includes "valgrind/valgrind.h" to link
  against the exposed "Valgrind::valgrind" library, so they can always
  find the header file.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20220428032855.64036-1-tchaikov@gmail.com>