util: add chunked_deque
chunked_deque provides an interface to a double-ended
queue, compatible with std::deque.

It uses chunked allocation very similar to
that of chunked_fifo, but its chunks and iterators
are bi-directional.

The motivation to add chunked_deque is mainly
to prevent large allocation seen with std::deque
(in libstdc++-11) that uses a single "spine" that
needes to be reallocated as a whole from time to time
when items are pushed into the container.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>