rpc_tester: Put client/server sides into correct sched groups
Right now the rpc slient is created in the default sched group.
Respectively, the server-side accepted connection is processed in the
single default sched group as well.

This patch makes each RPC job create its own client inside the properly
set scheduling group and negotiate the isolation cookie to the server.
On the server side the connection is isolated inside the peer sched
group as well.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220516145532.12442-1-xemul@scylladb.com>