Use tls socket to retrieve distinguished name
It's possible to receive DN information of the certificate used by the
peer using dn_calback. The callback is passed to the creadentials object
used on a server side. This callback is invoked during the handshake
(only if the 'client_auth' is set to REQUIRE) and accepts three parameters,
role (server or client), subject, and issuer.

The dn_callback is directly associated with credentials object. It can
be used to abort the handshake. But it's difficult to use the DN
information received by the callback for anything else.

This commit adds a 'get_dn_information' function to tls namespace.
This function returns a future that will be resolved with DN
information when the peer will send its leaf certificate. The function
accepts the connected_socket and forces TLS handshake. In case if client
certificate is not used it will return a nullopt.

This commit also adds a unit test and scripts that generate a bunch of
certificates/keys for it.

Signed-off-by: Evgeny Lazin <evgeny@vectorized.io>
Message-Id: <20220518150517.3537445-1-evgeny@redpanda.com>