build: use CMAKE_CXX_STANDARD instead of Seastar_CXX_DIALECT
before this change, Seastar_CXX_DIALECT is used to determine the
-std=<CXX_DIALECT> option passed to C++ compiler. to ensure the
corresponding CMAKE_CXX_STANDARD is enforced, we have to parse
the -std=<CXX_DIALECT> command line option. this is fragile, and
error-prone. as it is able to parse command line option like
-std=gnu++17, but would fail to parse -std=c++17, which is
a valid GCC option. moreoever, we should always respect the
CMAKE_CXX_STANDARD option specified by user instead of changing
it.

in this change, --c++-dialect option of configure.py is replaced
with --c++-standard. the latter option offloads the generation of
-std=<CXX_DIALECT> command line option to CMake. because CMake is
able to figure out the expected -std=<CXX_DIALECT> based on the
specified Seastar_CXX_STANDARD and the used compiler, we don't
need to worry about this anymore.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20220519134839.1818862-1-tchaikov@gmail.com>

[avi: adjust circleci configuration]