compat: do not use std::source_location if it is broken
in a recent clang change, __builtin_source_location was added.
see https://reviews.llvm.org/rGd61487490022 . but a bug in clang
frontend pratically prevents us from using it. see
https://github.com/llvm/llvm-project/issues/48230. so before this
issue is fixed, we have to prevent clang from using the builtin
source location implementation.

in this change, a new library is added for introducing the macro
of "SEASTAR_BROKEN_SOURCE_LOCATION", if it is defined we'd have to
use the homebrew implementation of source_location even the header
file of <source_location> or <experimental/source_location> can be
found by the C++ compiler.

please note, -std=c++20 is not added to CMAKE_REQUIRED_FLAGS when
performing the test, because we want to ensure that we are using
the same CXX_FLAGS as the ones when compiling std-compat.hh.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20220519134839.1818862-2-tchaikov@gmail.com>