rpc: remove recursion in do_unmarshall()
Currently, rpc uses recursion to process the argument list. This
means that for an argument list of size N, there are N calls to
std::tuple_cat, each with input tuple sizes in the range 0..N-1.
This is quadratic in N.

While the compiler *may* be able to keep optimize this away, this
is unlikely with non-trivial types. It's also likely no fun in terms
of compile time.

Fixing it is non-trivial since we must be prepared for an exception
while deserializing any of the arguments, and selectively destroying
only successfully constructed arguments. This is accomplished by
deserializing into a tuple of optional<T> instead of a tuple of T,
letting the engaged flag remember if we need to destroy the T or not.
A comma fold expression is used to deserialize while maintaining order.

We need to convert the tuple<optional<T>> into a tuple<T>, but that is
just linear in N, not quadratic.

The empty-tuple case (used to terminate recursion before) is converted
into a forward declaration, since do_unmarshall is used before it is
defined (by unmarhall_one, which is used by do_unmarshall...). Somehow
the compiler accepted the empty-tuple case even though it needed the
more general one before this patch.
Message-Id: <20220522165405.2612411-1-avi@scylladb.com>