json_elements: stream_range_as_array, stream_object: always close output stream
We added an assertion that the output_stream is
closed before it's destroyed to prevent use-after-free.
See 78b918ce6018a54d4e75e3da62c6313b50d71193

This means that the output stream must be closed after
it was written to, in all cases, successful or not.

Move the s.close() call to a finally clause so it would
always be called.

Failure to do so causes e.g. scylladb/scylla#10592

Refs scylladb/scylla#10592

(Refs rather than Fixes, as the fix will happen
when scylla updates its seastar submodule with this patch)

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>